feat: New page to create showcase posts (#556)
This was a bit more involved as the previous logic was for Github.
I also did not like some of the aspects of the form, including how we would upload images prior to submitting the full form, which would lead to images being mis-matched if people changed the slug after the facts.

This also includes a new input for selecting a FreeSewing user, and a different layout for the design input using buttons.

Reviewed-on: https://codeberg.org/freesewing/freesewing/pulls/556
Co-authored-by: Joost De Cock <joost@joost.at>
Co-committed-by: Joost De Cock <joost@joost.at>