[core] fix: Avoid duplicate intersections (#565)
Path.intersects sometimes returns the same intersection multiple times. Prevent this by checking for each new addition to the intersections array whether its distance to the final entry is larger than the tolerance. Tolerance is set (hard-coded) to 1e-3 mm, much less than the usual FreeSewing precision.

The duplicate intersections seem to be a result of the intersection occurring at a boundary between path segments, so that each segment contributes an entry.

Assumption: it is sufficient to compare only the last element of the existing array and the first of the new array.

`npm run test` inside the `packages/core` folder results in   416 passing (355ms), 1 pending

Co-authored-by: anna-puk <100537439+anna-puk@users.noreply.github.com>
Reviewed-on: https://codeberg.org/freesewing/freesewing/pulls/565
Reviewed-by: Joost De Cock <joostdecock@noreply.codeberg.org>
Co-authored-by: anna_puk <anna_puk@noreply.codeberg.org>
Co-committed-by: anna_puk <anna_puk@noreply.codeberg.org>